fix(operator): stop saving silent defaults for skipped hotel stars, distance, group type
The package wizard manufactured values for fields the operator skipped:
DEFAULT_DATA seeded distance='medium' and groupType='small-group', the
Zod schema re-applied .default('medium') on distance, and the Step 3
star select painted "4 stars" via ?? 4. A skipped field was therefore
stored as a confident default, as if the operator had confirmed it.

Now a skipped field persists as genuinely unset and renders "Not provided"
downstream (matching the display fix in PR #64):
- Stars: no painted default; select starts on a "Select hotel class"
  placeholder with an explicit "Not sure / not rated" option -> undefined.
- Distance: defaults to the existing 'unknown' sentinel (renders
  "Not provided"); "Not specified" option surfaced; no 'medium' coercion.
- Group type: no default; "Not specified" radio -> unset.

Zod schema extracted to lib/operator/package-schema.ts (unit-testable).
Adapter unchanged (already writes ?? null). Inclusions out of scope.
